requirements.txt
python-docx
tk